% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvhist.R
\name{asvhist}
\alias{asvhist}
\title{Plot an ASV read count histogram}
\usage{
asvhist(
  x,
  scale = "log",
  type = "freq",
  nbins = "FD",
  xlim = NULL,
  ylim = NULL,
  color = "black",
  fill = "grey"
)
}
\arguments{
\item{x}{A vector of ASV read counts (all positive numbers).}

\item{scale}{The scale of plot. Either \code{"log"} (common log, default) or
\code{"normal"}.}

\item{type}{The type of plot. Either \code{"freq"} or \code{"density"}; if \code{"freq"}
(default), frequencies (counts) are plotted; if \code{"density"}, probability
densities are plotted.}

\item{nbins}{Either a single number giving the number of bins or a character
string naming an algorithm to compute the number of bins (see ‘Details’).}

\item{xlim, ylim}{The ranges of x and y axes. Use to override the plot default.}

\item{color, fill}{Colors to be used to border and fill the bins. Use to
override the defaults (\code{"black"} and \code{"grey"}).}
}
\value{
A ggplot2 histogram plot
}
\description{
\code{asvhist()} creates a custom histogram plot of read counts of amplicon
sequence variants (ASVs) inferred from metabarcoding analysis.
}
\details{
The default for \code{nbins} is \code{"FD"}, which stands for
\code{"Freedman-Diaconis"}; see \code{\link[grDevices:nclass]{grDevices::nclass.FD()}}. Other options are
\code{"Scott"} and \code{"Sturges"}; see \code{\link[grDevices:nclass]{grDevices::nclass.scott()}} and
\code{\link[grDevices:nclass]{grDevices::nclass.Sturges()}}.
}
\examples{
data(mifish)
asvhist(mifish)
asvhist(mifish, type = "density", nbins = 30, xlim = c(1, 6))
}
