% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.gmmbs.R
\name{autoplot.gmmbs}
\alias{autoplot.gmmbs}
\title{Graphical representation of sequential parametric bootstrap tests for the
number of mixture components}
\usage{
\method{autoplot}{gmmbs}(object, title.size = NULL, annot.size = 6, ...)
}
\arguments{
\item{object}{A list object of class \code{gmmbs}.}

\item{title.size, annot.size}{The sizes of the plot title and annotation.
Use to override the default values.}

\item{...}{Unused.}
}
\value{
A ggplot2 plot
}
\description{
\code{autoplot.gmmbs()} displays a graphical representation of sequential
parametric bootstrap tests of k-component versus (k+1)-component
Gaussian mixture models, drawing a series of histogram plots. The
histogram in each plot shows the bootstrap distribution of likelihood
ratio statistic, to which the observed likelihood ratio (LR) is compared.
A vertical line signifying the observed LR is also drawn if it is not
extremely far from the bootstrap distribution.
}
\examples{
\dontrun{
  data(mifish)
  logmf <- log10(mifish)
  set.seed(101)
  # May take some time
  bs <- gmmbs(logmf, B = 100, epsilon = 1e-03)
  p <- autoplot(bs)
  library(cowplot)
  plot_grid(plotlist = p, ncol = 2)
  summary(bs)
}
}
\seealso{
\code{\link[=gmmbs]{gmmbs()}}
}
