% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.gmmcv.R
\name{autoplot.gmmcv}
\alias{autoplot.gmmcv}
\title{Graphical representation of cross-validation analysis for the number of
mixture components}
\usage{
\method{autoplot}{gmmcv}(object, ...)
}
\arguments{
\item{object}{A list object of class \code{gmmcv}.}

\item{...}{Unused.}
}
\value{
A ggplot2 plot
}
\description{
\code{autoplot.gmmcv()} displays a graphical representation of cross-validation
analysis for the number of components of Gaussian mixture model, drawing
the point plot of model's log-likelihood against the number of mixture
components, superimposed with the fitted LOESS curve and its confidence
interval.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
cv <- gmmcv(logmf, epsilon = 1e-03)
autoplot(cv)
}
\seealso{
\code{\link[=gmmcv]{gmmcv()}}
}
