% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.gmmem.R
\name{autoplot.gmmem}
\alias{autoplot.gmmem}
\title{Graphical representation of Gaussian mixture model fit to ASV read count data}
\usage{
\method{autoplot}{gmmem}(
  object,
  scale = "log",
  type = "freq",
  nbins = "FD",
  vline = NULL,
  xlim = NULL,
  ylim = NULL,
  hist.color = NULL,
  hist.fill = NULL,
  dist.color = NULL,
  dist.fill = NULL,
  dist.alpha = NULL,
  vline.color = NULL,
  ...
)
}
\arguments{
\item{object}{A list object of class \code{gmmem}.}

\item{scale}{The scale of plot. Either \code{"log"} (common log, default) or
\code{"normal"}.}

\item{type}{The type of plot. Either \code{"freq"} or \code{"density"}; if \code{"freq"}
(default), frequencies (counts) are plotted; if \code{"density"}, probability
densities are plotted.}

\item{nbins}{Either a single number giving the number of bins or a character
string naming an algorithm to compute the number of bins (see ‘Details’).}

\item{vline}{A numeric vector giving the quantiles of the mixture components.
Use to draw threshold(s) of the data. See the example.}

\item{xlim, ylim}{The ranges of x and y axes. Use to override the plot
defaults.}

\item{hist.color, hist.fill}{Colors to be used to border and fill the bins.
Use to override the defaults.}

\item{dist.color, dist.fill, dist.alpha}{The color (color and fill) and
opacity parameters for probability density functions. Use to override the
defaults.}

\item{vline.color}{The color for vertical lines.}

\item{...}{Unused.}
}
\value{
A ggplot2 plot
}
\description{
\code{autoplot.gmmem()} displays a graphical representation of k-component
Gaussian mixture model fit to ASV read count data, plotting the
probability density functions or count density functions of mixture
components over the histogram of log ASV read counts.
}
\details{
The default for \code{nbins} is \code{"FD"}, which stands for
\code{"Freedman-Diaconis"}; see \code{\link[grDevices:nclass]{grDevices::nclass.FD()}}. Other options are
\code{"Scott"} and \code{"Sturges"}; see \code{\link[grDevices:nclass]{grDevices::nclass.scott()}} and
\code{\link[grDevices:nclass]{grDevices::nclass.Sturges()}}, respectively.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
mod <- gmmem(logmf, k = 3)
autoplot(mod)
thresh <- quantile(mod, comp = 2)
autoplot(mod, vline = c(NA, thresh, NA))
}
\seealso{
\code{\link[=gmmem]{gmmem()}}, \code{\link[=predict.gmmem]{predict.gmmem()}}, \code{\link[=quantile.gmmem]{quantile.gmmem()}}
}
