% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.gmmem.R
\name{fitted.gmmem}
\alias{fitted.gmmem}
\title{Calculate probability density functions of mixture components}
\usage{
\method{fitted}{gmmem}(object, ...)
}
\arguments{
\item{object}{A list object of class \code{gmmem}.}

\item{...}{Unused.}
}
\value{
A data frame of class \code{fitted.gmmem} with three columns:
\describe{
\item{x}{The raw data.}
\item{comp}{Name of Mixture component.}
\item{density}{The probability densities.}
}
}
\description{
\code{fitted.gmmem()} calculates probability density functions (PDFs) of
individual mixture components of the fitted Gaussian mixture model.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
mod <- gmmem(logmf, k = 3)
pdf <- fitted(mod)
}
\seealso{
\code{\link[=gmmem]{gmmem()}}
}
