% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmbs.R
\name{gmmbs}
\alias{gmmbs}
\title{Perform sequential parametric bootstrap tests for the number of mixture
components}
\usage{
gmmbs(x, maxk = 10, B = 1000, sig = 0.05, epsilon = 1e-08, ...)
}
\arguments{
\item{x}{A vector of length n consisting of the data.}

\item{maxk}{The maximum number of components to consider.}

\item{B}{The number of bootstrap realizations of the likelihood ratio
statistic to produce. For details, see \code{\link[mixtools:boot.comp]{mixtools::boot.comp()}}.}

\item{sig}{The significance level for which to compare the p-value against
when performing the test of k-components versus (k+1)-components.}

\item{epsilon}{The convergence criterion. See \code{\link[mixtools:boot.comp]{mixtools::boot.comp()}}.}

\item{...}{Additional arguments passed to \code{\link[mixtools:boot.comp]{mixtools::boot.comp()}}.}
}
\value{
A list object of class \code{gmmbs} with items:
\describe{
\item{p.values}{The p-values for each test of k-components vs.
(k+1)-components.}
\item{log.lik}{The B bootstrap realizations of the likelihood ratio
statistic for each test.}
\item{obs.log.lik}{The observed likelihood ratio statistic for each test
which is used in determining the p-values.}
}
}
\description{
\code{gmmbs()} is a simple wrapper of \code{\link[mixtools:boot.comp]{mixtools::boot.comp()}} to perform
sequential parametric bootstrap tests of k-component versus
(k+1)-component Gaussian mixture models for selecting the number of
components.
}
\examples{
\dontrun{
  data(mifish)
  logmf <- log10(mifish)
  set.seed(101)
  # May take some time
  bs <- gmmbs(logmf, B = 100, epsilon = 1e-03)
  p <- autoplot(bs)
  library(cowplot)
  plot_grid(plotlist = p, ncol = 2)
  summary(bs)
}
}
\references{
\itemize{
\item Benaglia, T., Chauveau, D., Hunter, D. R., & Young, D. S. (2009). mixtools:
An R package for analyzing finite mixture models. Journal of Statistical
Software, 32(6), 1-29. \url{https://doi.org/10.18637/jss.v032.i06}
\item Shalizi, C. R. (2011). Lecture 20, Mixture Examples and Complements.
In 36-402, Advanced Data Analysis (pp. 1-23).
\url{https://www.stat.cmu.edu/~cshalizi/402/lectures/20-mixture-examples/lecture-20.pdf}
}
}
