% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmcv.R
\name{gmmcv}
\alias{gmmcv}
\title{Perform cross-validation analysis for the number of mixture components}
\usage{
gmmcv(x, maxk = 10, maxit = 1000, maxrestarts = 20, epsilon = 1e-08)
}
\arguments{
\item{x}{A vector of length n consisting of the data.}

\item{maxk}{The maximum number of components to consider. Default is 10.}

\item{maxit}{The maximum number of iterations. Default is 1000.}

\item{maxrestarts}{The maximum number of restarts allowed in case of a
problem with the particular starting values chosen due to one of the
variance estimates getting too small (each restart uses randomly chosen
starting values). Default is 20.}

\item{epsilon}{The convergence criterion. Default is 1e-08.}
}
\value{
A list object of class \code{gmmcv} with items:
\describe{
\item{k}{The number of mixture components considered.}
\item{log.lik}{The log-likelihood of the k-component Gaussian mixture model.}
}
}
\description{
\code{gmmcv()} performs split-half cross-validation analysis for selecting the
number of components of Gaussian mixture model, using an EM algorithm
implemented by \code{\link[mixtools:normalmixEM]{mixtools::normalmixEM()}} to estimate model parameters.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
cv <- gmmcv(logmf, epsilon = 1e-03)
autoplot(cv)
}
\references{
\itemize{
\item Benaglia, T., Chauveau, D., Hunter, D. R., & Young, D. S. (2009). mixtools:
An R package for analyzing finite mixture models. Journal of Statistical
Software, 32(6), 1-29. \url{https://doi.org/10.18637/jss.v032.i06}
\item Shalizi, C. R. (2011). Lecture 20, Mixture Examples and Complements.
In 36-402, Advanced Data Analysis (pp. 1-23).
\url{https://www.stat.cmu.edu/~cshalizi/402/lectures/20-mixture-examples/lecture-20.pdf}
}
}
\seealso{
\code{\link[mixtools:normalmixEM]{mixtools::normalmixEM()}}
}
