% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gmmem.R
\name{predict.gmmem}
\alias{predict.gmmem}
\title{Extract predicted values from PDFs of mixture components}
\usage{
\method{predict}{gmmem}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A list object of class \code{gmmem}.}

\item{newdata}{An optional vector of length n consisting of the new data. If
NULL (default), the fitted values are used.}

\item{...}{Unused.}
}
\value{
A data frame of class \code{predict.gmmem} with three columns:
\describe{
\item{x}{The data to which probability densities are predicted.}
\item{comp}{Name of Mixture component.}
\item{density}{The probability densities.}
}
}
\description{
\code{predict.gmmem()} extracts predicted values from probability density
functions (PDFs) of individual mixture components of the fitted Gaussian
mixture model.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
mod <- gmmem(logmf, k = 3)
pdf <- predict(mod)
}
\seealso{
\code{\link[=gmmem]{gmmem()}}
}
