% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gmmbs.R
\name{summary.gmmbs}
\alias{summary.gmmbs}
\title{Summarize the result of sequential parametric bootstrap tests for the number
of mixture components}
\usage{
\method{summary}{gmmbs}(object, ...)
}
\arguments{
\item{object}{A data frame object of class \code{gmmbs}.}

\item{...}{Unused.}
}
\value{
A data frame object with two columns:
\describe{
\item{LR}{The observed likelihood ratio statistic for each test of
k-components versus (k+1)-components.}
\item{p.val}{The p-value for each test.}
}
}
\description{
\code{summary.gmmbs()} creates a data frame object of the summary statistics of
sequential parametric bootstrap tests for the number of mixture
components.
}
\examples{
\dontrun{
  data(mifish)
  logmf <- log10(mifish)
  set.seed(101)
  # May take some time
  bs <- gmmbs(logmf, B = 100, epsilon = 1e-03)
  p <- autoplot(bs)
  library(cowplot)
  plot_grid(plotlist = p, ncol = 2)
  summary(bs)
}
}
\seealso{
\code{\link[=gmmbs]{gmmbs()}}
}
