% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmDenoise-package.R
\docType{package}
\name{gmmDenoise-package}
\alias{gmmDenoise}
\alias{gmmDenoise-package}
\title{gmmDenoise: A set of functions for filtering erroneous sequences in eDNA metabarcoding data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions for filtering erroneous sequences or amplicon sequence variants (ASVs) in eDNA metabarcoding data, based on Gaussian mixture modeling (GMM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YSKoseki/gmmDenoise}
}

}
\author{
\strong{Maintainer}: Yusuke Koseki \email{ykoseki@otsuma.ac.jp} (\href{https://orcid.org/0000-0001-9588-3018}{ORCID})

}
\keyword{internal}
