% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.gmmem.R
\name{quantile.gmmem}
\alias{quantile.gmmem}
\title{Calculate quantiles of a mixture component}
\usage{
\method{quantile}{gmmem}(x, comp = 2, probs = 0.95, lower.tail = TRUE, ...)
}
\arguments{
\item{x}{A list object of class \code{gmmem}.}

\item{comp}{A numeric vector specifying the component numbers. Note that the
component numbers are in ascending order from the uppermost component.}

\item{probs}{A vector of probabilities. Elements must be [0 ≤ probs ≤ 1].}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
P[X ≤ x] otherwise, P[X > x].}

\item{...}{Unused.}
}
\value{
A vector of quantiles.
}
\description{
\code{quantile.gmmem()} calculates quantiles of mixture components (default is
the upper 95\% of the second uppermost component) of the fitted Gaussian
mixture model. The quantile calculated can be referred to as the read
count cutoff threshold for amplicon sequence variant (ASV) filtering.
}
\examples{
data(mifish)
logmf <- log10(mifish)
set.seed(101)
mod <- gmmem(logmf, k = 3)
autoplot(mod)
thresh <- quantile(mod)
autoplot(mod, vline = thresh)
}
\seealso{
\link{gmmem}
}
